replace covar with predictions
